chore: Update Tailwind CSS configuration and styles
This commit updates the Tailwind CSS configuration file `tailwind.config.js` to include the `darkMode` option set to "selector". It also adds new styles to the `output.css` file, including the `.text-4xl` and `.text-blue-500` classes. These styles are used to adjust font sizes and colors in the templates. Additionally, the `home.html` and `about.html` templates have been modified to use the `fragments/base.html` template as the parent template, providing a consistent structure for the pages.