chore: Update base.html and footer.html templates
This commit updates the `base.html` template by commenting out the footer section to temporarily remove it from the page. This change is made to address a bug related to the footer's positioning. Additionally, the `footer.html` template is modified to include the class `mt-10` for proper spacing. These changes improve the overall layout and user experience of the website.