chore: Update PageController with sign-up and login functionality
This commit adds sign-up and login functionality to the PageController. The `signUp` method handles the GET request for the "/sign-up" endpoint and sets up the necessary model attributes for the "signUp" template. The `doSignUp` method handles the POST request for the "/do-signUp" endpoint and saves the user form data using the UserService. The `login` method handles the GET request for the "/login" endpoint and sets up the necessary model attributes for the "login" template. These changes enhance the user experience by allowing users to sign up and log in to the Smart Contact Manager application.